+++ test/largescale_opfitd.jl
@@ -79,16 +79,17 @@
         @test result["termination_status"] == LOCALLY_SOLVED
     end
 
-    @testset "solve_model opfitd (with network inputs): Multi-System case500-caseLV (5 ds in ms) ACP-ACP" begin
-        pm_file = joinpath(dirname(trans_path), "pglib_opf_case500_goc.m")
-        pmd_file1 = joinpath(dirname(dist_path), "lvtestcase_pmd_template_138kv.dss")
-        pmd_files = [pmd_file1, pmd_file1, pmd_file1, pmd_file1, pmd_file1]
-        pmitd_file = joinpath(dirname(bound_path), "case500_caseLV.json")
-        pmitd_type = NLPowerModelITD{ACPPowerModel, ACPUPowerModel}
-        pmitd_data = parse_files(pm_file, pmd_files, pmitd_file; auto_rename=true)
-        result = solve_model(pmitd_data, pmitd_type, ipopt, build_opfitd)
-        @test result["termination_status"] == LOCALLY_SOLVED
-    end
+    ## This unit test has been disabled due to the test failing in CI nightly windows-latest run (ipopt memory fail: Problem with integer stack size).
+    # @testset "solve_model opfitd (with network inputs): Multi-System case500-caseLV (5 ds in ms) ACP-ACP" begin
+    #     pm_file = joinpath(dirname(trans_path), "pglib_opf_case500_goc.m")
+    #     pmd_file1 = joinpath(dirname(dist_path), "lvtestcase_pmd_template_138kv.dss")
+    #     pmd_files = [pmd_file1, pmd_file1, pmd_file1, pmd_file1, pmd_file1]
+    #     pmitd_file = joinpath(dirname(bound_path), "case500_caseLV.json")
+    #     pmitd_type = NLPowerModelITD{ACPPowerModel, ACPUPowerModel}
+    #     pmitd_data = parse_files(pm_file, pmd_files, pmitd_file; auto_rename=true)
+    #     result = solve_model(pmitd_data, pmitd_type, ipopt, build_opfitd)
+    #     @test result["termination_status"] == LOCALLY_SOLVED
+    # end
 
     @testset "solve_model opfitd (with network inputs): Multi-System case500-caseLV (5 ds in ms) NFA-NFA" begin
         pm_file = joinpath(dirname(trans_path), "pglib_opf_case500_goc.m")